+++ src/main/resources/application.properties
@@ -5,3 +5,6 @@ spring.datasource.password=javabootcamp
 
 #drop n create table again, good for testing, comment this in production
 spring.jpa.hibernate.ddl-auto=create-drop
+
+server.servlet.context-path=/simple-blog
+server.port=8081
